refactor: reorganize source structure
