+++ src/components/Formula.jsx
