+++ src/hooks/index.js
@@ -1,0 +1,2 @@
+export { default as useLineData } from './useLineData';
+export { default as useSedimentData } from './useSedimentData';
