+++ src/pages/index.js
@@ -1,0 +1,2 @@
+export { default as Help } from './Help';
+export { default as Normativa } from './Normativa';
