+++ src/styles/Widget.css
